pyproject.toml
[tool.poetry]
name = "coc-api"
version = "0.1.0"
description = ""
authors = ["yOung-4 <91882244+yOung-4@users.noreply.github.com>"]
readme = "README.md"
packages = [{include = "coc_api"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.89.1"
python-dotenv = "^1.0.0"
pony = "^0.7.16"
psycopg2-binary = "^2.9.6"
passlib = "^1.7.4"
bcrypt = "^4.0.1"
union = "^0.1.10"
python-jose = "^3.3.0"
cryptography = "^41.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
